#1270 Clean up ported Csla.Xaml code
